Conflicts:
	src/plugins/common/commands/lib/inserthtml.js
	src/test/unit/commandtest.js

+++ bin/deploy-aloha.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
-# $Id: deploy_aloha, v 1.0 2011/10/3 15:50:00 jotschi Exp $
+# $Id: deploy-aloha, v 1.0 2011/10/3 15:50:00 jotschi Exp $
 # 2011 - written by Johannes Schüth <j.schueth@gentics.com>
 
-SCRIPT="`readlink -f $0`" 
-BASEDIR="`dirname "$SCRIPT"`"
+BASEDIR="$( cd "$( dirname "$0" )" && pwd )"
 
 # Deployment url
 DURL=$1
 DPATH=$2
+
+if [ "$DURL" == "NA" ] || [ "$DURL" == "SKIP" ] ; then
+  echo "No valid deployment url specified. I'll silently omitt deployment via ssh"
+  exit 0 
+fi
 
 echo -e "\n * Removing old archives from $DPATH"
   ssh $DURL "cd $DPATH ; rm -rf *"
